chat_app/requirements.txt
Flask==3.0.3
gunicorn==23.0.0
numpy==1.24.3
pandas==2.1.1
requests
python-dotenv==1.0.0
swarm @ git+https://github.com/openai/swarm.git
